+++ kernax/multioutput/__init__.py
@@ -2,12 +2,12 @@ from .BlockDiagKernel import BlockDiagKernel
 from .BlockMean import BlockMean
 from .ConvolutionKernel import ConvolutionKernel
 from .ICMKernel import ICMKernel
-from .LMCKernel import LMCKernel
+from .LCMKernel import LCMKernel
 
 __all__ = [
 	"BlockDiagKernel",
 	"BlockMean",
 	"ConvolutionKernel",
 	"ICMKernel",
-	"LMCKernel",
+	"LCMKernel",
 ]